Revert "Feature/i18n integration (#1)"
This reverts commit e413e003772c4b0c48d03a1b5d7973f024434548.
