--- next.config.mjs
@@ -1,15 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  // Configuración de redirecciones explícitas
-  async redirects() {
-    return [
-      {
-        source: '/',
-        destination: '/es',
-        permanent: true,
-      },
-    ];
-  },
-};
-
-export default nextConfig;
