+++ package.json
@@ -1,16 +1,15 @@
 {
   "name": "poke-cool",
-  "version": "0.2.0",
+  "version": "0.1.1",
   "private": true,
   "description": "A Next.js practice project. Made by Juan Felipe Parra Runceria.",
   "scripts": {
-    "dev": "node scripts/generate-translation-types.js && next dev --turbopack",
-    "build": "node scripts/generate-translation-types.js && next build",
+    "dev": "next dev --turbopack",
+    "build": "next build",
     "start": "next start",
     "lint": "next lint",
     "format": "prettier --write .",
-    "format:check": "prettier --check .",
-    "generate-types": "node scripts/generate-translation-types.js"
+    "format:check": "prettier --check ."
   },
   "dependencies": {
     "@radix-ui/react-slot": "^1.1.2",
@@ -20,7 +19,6 @@
     "framer-motion": "^12.5.0",
     "lucide-react": "^0.482.0",
     "next": "15.2.3",
-    "next-international": "^1.3.1",
     "next-themes": "^0.4.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
