--- src/app/components/LanguageSwitcher/translations.json
@@ -1,12 +1,0 @@
-{
-  "es": {
-    "change_language": "Cambiar idioma",
-    "language_es": "Español",
-    "language_en": "Inglés"
-  },
-  "en": {
-    "change_language": "Change language",
-    "language_es": "Spanish",
-    "language_en": "English"
-  }
-}
