--- src/app/components/ServerComponent/translations.json
@@ -1,10 +1,0 @@
-{
-  "es": {
-    "title": "Hola mundo",
-    "description": "Este es un componente traducido"
-  },
-  "en": {
-    "title": "Hello world",
-    "description": "This is a translated component"
-  }
-}
