--- src/app/components/ThemeToggle/translations.json
@@ -1,14 +1,0 @@
-{
-  "es": {
-    "light": "Modo claro",
-    "dark": "Modo oscuro",
-    "system": "Sistema",
-    "toggle_theme": "Cambiar tema"
-  },
-  "en": {
-    "light": "Light mode",
-    "dark": "Dark mode",
-    "system": "System",
-    "toggle_theme": "Toggle theme"
-  }
-}
