+++ src/app/layout.tsx
@@ -1,10 +1,8 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
-import '../globals.css';
-import { ThemeProvider } from '@/app/providers/ThemeProvider';
-import { notFound } from 'next/navigation';
+import './globals.css';
+import { ThemeProvider } from '@/app/components/providers/ThemeProvider';
 
-// Configuración de las fuentes
 const geistSans = Geist({
   variable: '--font-geist-sans',
   subsets: ['latin'],
@@ -16,28 +14,17 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: 'PokeCool',
-  description: 'Your cool and online Pokédex',
+  title: 'Create Next App',
+  description: 'Generated by create next app',
 };
 
-export function generateStaticParams() {
-  return [{ locale: 'es' }, { locale: 'en' }];
-}
-
-export default async function RootLayout({
-  params,
+export default function RootLayout({
   children,
-}: {
-  params: Promise<{ locale: string }>;
+}: Readonly<{
   children: React.ReactNode;
-}) {
-  const { locale } = await params;
-  if (locale !== 'es' && locale !== 'en') {
-    notFound();
-  }
-
+}>) {
   return (
-    <html lang={locale} suppressHydrationWarning>
+    <html lang="en" suppressHydrationWarning>
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
