+++ src/app/page.tsx
@@ -1,48 +1,30 @@
 import Image from 'next/image';
-import { ThemeToggle } from '../components/ThemeToggle';
-import LanguageSwitcher from '../components/LanguageSwitcher';
-import { getGlobalTranslations } from '@/utils';
-import { GlobalMessage } from '../components/GlobalMessage';
-import ServerComponent from '../components/ServerComponent';
-import { getCurrentLocale } from '@/utils/getCurrentLocale';
+import { ThemeToggle } from './components/ThemeToggle';
 
-export default async function Home() {
-  const locale = await getCurrentLocale();
-  const t = getGlobalTranslations(locale);
-
+export default function Home() {
   return (
     <div className="grid min-h-screen grid-rows-[20px_1fr_20px] items-center justify-items-center gap-16 p-8 pb-20 font-[family-name:var(--font-geist-sans)] sm:p-20">
       <main className="row-start-2 flex flex-col items-center gap-[32px] sm:items-start">
-        <div className="flex w-full items-center justify-between">
-          <Image
-            className="dark:invert"
-            src="/next.svg"
-            alt="Next.js logo"
-            width={180}
-            height={38}
-            priority
-          />
-          <div className="flex items-center gap-4">
-            <ThemeToggle />
-            <LanguageSwitcher />
-          </div>
-        </div>
-
-        <div className="w-full max-w-md">
-          <GlobalMessage />
-          <ServerComponent />
-        </div>
-
+        <Image
+          className="dark:invert"
+          src="/next.svg"
+          alt="Next.js logo"
+          width={180}
+          height={38}
+          priority
+        />
+        <ThemeToggle />
         <ol className="list-inside list-decimal text-center font-[family-name:var(--font-geist-mono)] text-sm/6 sm:text-left">
           <li className="mb-2 tracking-[-.01em]">
-            {t.messages.get_started}{' '}
+            Get started by editing{' '}
             <code className="rounded bg-black/[.05] px-1 py-0.5 font-[family-name:var(--font-geist-mono)] font-semibold dark:bg-white/[.06]">
               src/app/page.tsx
             </code>
             .
           </li>
-          <li className="tracking-[-.01em]">{t.messages.hello}</li>
-          <li className="tracking-[-.01em]">{t.messages.welcome}</li>
+          <li className="tracking-[-.01em]">
+            Save and see your changes instantly.
+          </li>
         </ol>
 
         <div className="flex flex-col items-center gap-4 sm:flex-row">
@@ -59,7 +41,7 @@ export default async function Home() {
               width={20}
               height={20}
             />
-            {t.messages.deploy_now}
+            Deploy now
           </a>
           <a
             className="flex h-10 w-full items-center justify-center rounded-full border border-solid border-black/[.08] px-4 text-sm font-medium transition-colors hover:border-transparent hover:bg-[#f2f2f2] sm:h-12 sm:w-auto sm:px-5 sm:text-base md:w-[158px] dark:border-white/[.145] dark:hover:bg-[#1a1a1a]"
@@ -67,7 +49,7 @@ export default async function Home() {
             target="_blank"
             rel="noopener noreferrer"
           >
-            {t.messages.read_docs}
+            Read our docs
           </a>
         </div>
       </main>
@@ -85,7 +67,7 @@ export default async function Home() {
             width={16}
             height={16}
           />
-          {t.messages.learn}
+          Learn
         </a>
         <a
           className="flex items-center gap-2 hover:underline hover:underline-offset-4"
@@ -100,7 +82,7 @@ export default async function Home() {
             width={16}
             height={16}
           />
-          {t.messages.examples}
+          Examples
         </a>
         <a
           className="flex items-center gap-2 hover:underline hover:underline-offset-4"
@@ -115,7 +97,7 @@ export default async function Home() {
             width={16}
             height={16}
           />
-          {t.messages.go_to_nextjs}
+          Go to nextjs.org →
         </a>
       </footer>
     </div>
