--- src/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useComponentTranslations';
-export * from './useLocale';
-export * from './useGlobalTranslations';
