--- src/i18n/common.json
@@ -1,34 +1,0 @@
-{
-  "es": {
-    "messages": {
-      "welcome": "Bienvenido a la Pokédex",
-      "hello": "Hola",
-      "get_started": "Comienza editando",
-      "deploy_now": "Desplegar ahora",
-      "read_docs": "Leer documentación",
-      "learn": "Aprender",
-      "examples": "Ejemplos",
-      "go_to_nextjs": "Ir a nextjs.org →"
-    },
-    "languages": {
-      "es": "ES",
-      "en": "EN"
-    }
-  },
-  "en": {
-    "messages": {
-      "welcome": "Welcome to the Pokédex",
-      "hello": "Hello",
-      "get_started": "Get started by editing",
-      "deploy_now": "Deploy now",
-      "read_docs": "Read our docs",
-      "learn": "Learn",
-      "examples": "Examples",
-      "go_to_nextjs": "Go to nextjs.org →"
-    },
-    "languages": {
-      "es": "ES",
-      "en": "EN"
-    }
-  }
-}
